drivers: fwtable: update board type to be uint32_t
Update board type to uint32_t to support up to five hex digits. Also
update galaxy bin6 board type to accurately represent type (0x202).

+++ lib/tenstorrent/bh_arc/chip_info_fwtable.c
@@ -26,7 +26,7 @@ BUILD_ASSERT((int)BH_PCIE_MODE_RP == FwTable_PciPropertyTable_PcieMode_RP);
 
 bool bh_chip_info_is_ubb(void)
 {
-	uint8_t board_type = tt_bh_fwtable_get_board_type(fwtable_dev);
+	uint32_t board_type = tt_bh_fwtable_get_board_type(fwtable_dev);
 
 	return board_type == BOARDTYPE_UBB || board_type == BOARDTYPE_GALAXY_BIN6;
 }
